Ensure that orocos_kdl_vendor doesn't accidentally find itself.
When initially building the orocos_kdl_vendor package (on platforms
where it actually builds), it turns out that it places a
valid cmake configuration in the build directory.  In turn,
that means that a subsequent rebuild will find this configuration
in the build directory, and throw the rest of the logic off.
This only seems to be a problem with CMake 3.29 and later, though
I can't say exactly why at the moment.

Workaround this problem by writing the configuration out to a
temporary file, and then moving it into the final place with
the final name.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>